+++ R/r2_coxsnell.R
@@ -218,25 +218,20 @@ r2_coxsnell.svycoxph <- function(model, ...) {
 
 #' @export
 r2_coxsnell.multinom <- function(model, ...) {
-  l_base <- insight::get_loglikelihood(stats::update(model, ~1, trace = FALSE))
+  l_base <- insight::get_loglikelihood(insight::null_model(model))
   .r2_coxsnell(model, l_base)
 }
 
 #' @export
-r2_coxsnell.clm2 <- function(model, ...) {
-  l_base <- insight::get_loglikelihood(stats::update(model, location = ~1, scale = ~1))
-  .r2_coxsnell(model, l_base)
-}
+r2_coxsnell.clm2 <- r2_coxsnell.multinom
 
 #' @export
-r2_coxsnell.bayesx <- function(model, ...) {
-  junk <- utils::capture.output(l_base <- insight::get_loglikelihood(stats::update(model, ~1))) # nolint
-  .r2_coxsnell(model, l_base)
-}
+r2_coxsnell.bayesx <- r2_coxsnell.multinom
 
 #' @export
 r2_coxsnell.clm <- function(model, ...) {
-  l_base <- insight::get_loglikelihood(stats::update(model, ~1))
+  l_base <- insight::get_loglikelihood(insight::null_model(model))
+
   # if no loglik, return NA
   if (length(as.numeric(l_base)) == 0) {
     return(NULL)