+++ R/r2_mcfadden.R
@@ -73,7 +73,7 @@ r2_mcfadden.glm <- function(model, verbose = TRUE, ...) {
     return(NULL)
   }
 
-  l_null <- insight::get_loglikelihood(stats::update(model, ~1))
+  l_null <- insight::get_loglikelihood(insight::null_model(model))
   .r2_mcfadden(model, l_null)
 }
 
@@ -162,23 +162,20 @@ r2_mcfadden.vglm <- function(model, ...) {
     insight::format_error("Can't get log-likelihood when `summ` is not zero.")
   }
 
-  l_null <- insight::get_loglikelihood(stats::update(model, ~1))
+  l_null <- insight::get_loglikelihood(insight::null_model(model))
   .r2_mcfadden(model, l_null)
 }
 
 
 #' @export
 r2_mcfadden.clm2 <- function(model, ...) {
-  l_null <- insight::get_loglikelihood(stats::update(model, location = ~1, scale = ~1))
+  l_null <- insight::get_loglikelihood(insight::null_model(model))
   .r2_mcfadden(model, l_null)
 }
 
 
 #' @export
-r2_mcfadden.multinom <- function(model, ...) {
-  l_null <- insight::get_loglikelihood(stats::update(model, ~1, trace = FALSE))
-  .r2_mcfadden(model, l_null)
-}
+r2_mcfadden.multinom <- r2_mcfadden.clm2
 
 
 #' @export