+++ NEWS.md
@@ -6,6 +6,9 @@
   functions provide information about the reliability of group-level estimates
   (i.e., random effects) in mixed models.
 
+* `check_group_variation()` to check whether variables have a within- or between-
+  group variation (or both).
+
 ## Changes
 
 * Singularity checks with `check_singularity()` are now more efficient and also