+++ NEWS.md
@@ -1,4 +1,4 @@
-# performance (devel)
+# performance 0.14.0
 
 ## Breaking Changes